Add selection of Salesforce environment type
Users should also be allowed to select the type of Salesforce environment. I could not see any parameters in the Cloudformation template, hence I have added it here.

Also note though the "sf_is_production" environment variable has been added to the "AWSSCV_vmx_packager_*" function, still the function is running into error. The reason to this is that "Mode" variable in "sf_config.py" file in the python layer has been set to "Mode", the value of which is picked from "awsscv_salesforce_config_*" entry in secret manager. I could not find any variable with the name "Mode" in this this entry. Hence I think the there should be an addition of "Mode" variable to the "awsscv_salesforce_config_*" secret entry while deploying the Voicemail solution.